Maintain branch name semantics in git clone
When you checkout a SHA as it's done here, git goes into detached head mode and loses all the nice metadata related to the branch. Ran into an issue with this when running code coverage tooling which would upload its results and tag it using git metadata.

This achieves the exact same purpose of checking out a particular SHA (note that the clone is hard-reset to the SHA on L30, but within the context of the checked out branch), and maintains the same printing, but is a little cleaner.